# std: util to encode TLV to string