--- a/std/utils/testutils/error.go
+++ b/std/utils/testutils/error.go
@@ -21,3 +21,17 @@ func Err[T any](_ T, err error) error {
 	require.Error(testT, err)
 	return err
 }
+
+func NoErrB[T any](v T, err error) T {
+	if err != nil {
+		panic(err)
+	}
+	return v
+}
+
+func ErrB[T any](_ T, err error) error {
+	if err == nil {
+		panic("expected error")
+	}
+	return err
+}