Better version of #121

> enc.Name.String() is too slow for usage in map[string]... (since go doesn't support custom keys). Using a cryptographic hash is too expensive and xxhash is not collision free.

`TlvStr()` returns a TLV encoded name string. I realised (bit late) that our encoding is very efficient. Decoding is slightly slower than the earlier version but that's fine since map reads need only encode.

Surprisingly the TLV encode is even faster than just copying the memory over (huh go?)

```
# /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkName github.com/named-data/ndnd/std/encoding
goos: linux
goarch: amd64
pkg: github.com/named-data/ndnd/std/encoding
cpu: AMD EPYC 7702P 64-Core Processor               
BenchmarkNameHash-128                            1000000              1028 ns/op             160 B/op         20 allocs/op
BenchmarkNameStringEncode-128                     112825             10550 ns/op            3467 B/op         30 allocs/op
BenchmarkNameTlvStrEncode-128                    2813996               415.0 ns/op           448 B/op          1 allocs/op
BenchmarkNameBytesEncode-128                     2309168               436.1 ns/op           448 B/op          1 allocs/op
BenchmarkNameComponentStringEncode-128           2747901               468.1 ns/op            95 B/op          4 allocs/op
BenchmarkNameComponentTlvStrEncode-128          24632506                46.90 ns/op           16 B/op          1 allocs/op
BenchmarkNameStringDecode-128                      94125             12469 ns/op            2059 B/op         22 allocs/op
BenchmarkNameTlvStrDecode-128                     575214              2039 ns/op            2400 B/op          5 allocs/op
BenchmarkNameComponentStringDecode-128           4044622               290.7 ns/op            23 B/op          1 allocs/op
BenchmarkNameComponentTlvStrDecode-128           8897691               131.1 ns/op            48 B/op          2 allocs/op
PASS
ok      github.com/named-data/ndnd/std/encoding 244.251s
```

After moving `Name.Hash()` to TLV (reduce alloc):
```
BenchmarkNameHash-128            2060044               528.0 ns/op           448 B/op          1 allocs/op
```

Same for `Name.Clone()`
```
Before -
BenchmarkNameClone-128            800976              1820 ns/op            1128 B/op         21 allocs/op
After -
BenchmarkNameClone-128           1503552               735.4 ns/op          1056 B/op          2 allocs/op
```